+++ docs/sphinx/conf.py
@@ -188,4 +188,3 @@ texinfo_documents = [
 ]
 
 
-