+++ examples/size_limiter.cpp
@@ -7,7 +7,7 @@
 #include <iostream>
 
 #include "umpire/ResourceManager.hpp"
-#include "umpire/strategy/DynamicPool.hpp"
+#include "umpire/strategy/QuickPool.hpp"
 #include "umpire/strategy/SizeLimiter.hpp"
 #include "umpire/util/Macros.hpp"
 
@@ -17,7 +17,7 @@ int main(int, char**)
   auto size_limited_alloc = rm.makeAllocator<umpire::strategy::SizeLimiter>(
       "size_limited_alloc", rm.getAllocator("HOST"), 1024);
 
-  auto pool = rm.makeAllocator<umpire::strategy::DynamicPool>(
+  auto pool = rm.makeAllocator<umpire::strategy::QuickPool>(
       "pool", size_limited_alloc, 64, 64);
 
   // This will throw an exception because the pool is limited to 1024 bytes.