+++ examples/use_after_free.cpp
@@ -7,13 +7,13 @@
 #include <iostream>
 
 #include "umpire/ResourceManager.hpp"
-#include "umpire/strategy/DynamicPoolList.hpp"
-#include "umpire/strategy/DynamicPoolMap.hpp"
+#include "umpire/strategy/QuickPool.hpp"
+#include "umpire/strategy/QuickPool.hpp"
 
 int main(int, char**)
 {
   auto& rm = umpire::ResourceManager::getInstance();
-  auto pool = rm.makeAllocator<umpire::strategy::DynamicPoolMap>(
+  auto pool = rm.makeAllocator<umpire::strategy::QuickPool>(
       "pool", rm.getAllocator("HOST"));
 
   double* data = static_cast<double*>(pool.allocate(1024 * sizeof(double)));