+++ src/umpire/Replay.hpp
@@ -12,11 +12,6 @@
 #include <sstream>
 #include <string>
 
-#include "umpire/strategy/DynamicPool.hpp"
-#include "umpire/strategy/DynamicPoolList.hpp"
-#include "umpire/strategy/DynamicPoolMap.hpp"
-#include "umpire/strategy/QuickPool.hpp"
-
 namespace umpire {
 
 namespace {
@@ -26,13 +21,6 @@ int m_argument_number;
 class Allocator;
 
 std::ostream& operator<<(std::ostream& out, umpire::Allocator&);
-
-std::ostream& operator<<(std::ostream& out,
-                         umpire::strategy::DynamicPoolMap::CoalesceHeuristic&);
-std::ostream& operator<<(std::ostream& out,
-                         umpire::strategy::DynamicPoolList::CoalesceHeuristic&);
-std::ostream& operator<<(std::ostream& out,
-                         umpire::strategy::QuickPool::CoalesceHeuristic&);
 
 class Replay {
  public: